Write symbolic refs
In a previous change we implemented the ability to write to a symbolic
ref, but we could only write OIDs to it. Now we implement the ability to
write a symbolic value to a ref (and it doesn't matter if the ref that
we write is itself symbolic or non-symbolic).