Fix log crash
This started out with rewriting the data::get_ref_internal() logic to
more closely match the tutorial. While possibly not essential, it's
probably a good thing for future updates.

The actual bug seemed to be caused by a missing call to data::get_ref()
on line 26 of base.rs. When fixed, this exposed an infinite recursion
bug caused by passing the wrong value to data::get_ref_internal() on
line 99 of data.rs.