Download remote ref values
Instead of just printing remote refs, let's actually save them locally.
We're going to use a separate ref namespace for remote refs, so if the
remote repository has refs/heads/master we're going to save it locally
as refs/remote/master to distinguish the remote master from the local
master.